chore: move non-shared code from frontend-shared to app (#24674)
* move non shared code to packages/app

* move things around

* fix bug

* remove unused code

* remove unused code

* revert change

* unused

* fix code

* revert [skip ci]

* revert [skip ci]
