+++ packages/app/src/specs/generators/scaffold/ScaffoldGeneratorStepOne.vue
@@ -58,8 +58,8 @@
 import { gql, useMutation } from '@urql/vue'
 import { ScaffoldGeneratorStepOne_ScaffoldIntegrationDocument } from '../../../generated/graphql'
 import { computed, onMounted } from 'vue'
-import { buildSpecTree } from '../../spec-utils'
-import { useCollapsibleTree } from '@packages/frontend-shared/src/composables/useCollapsibleTree'
+import { buildSpecTree } from '../../../specs/tree/useCollapsibleTree'
+import { useCollapsibleTree } from '../../tree/useCollapsibleTree'
 import StandardModalFooter from '@cy/components/StandardModalFooter.vue'
 import Button from '@cy/components/Button.vue'
 import { useI18n } from '@cy/i18n'
