--- packages/frontend-shared/src/composables/examples/UseCollapsibleTreeExample.cy.tsx
@@ -1,7 +1,0 @@
-import UseCollapsibleTreeExample from './UseCollapsibleTreeExample.vue'
-
-describe('<UseCollapsibleTreeExample />', () => {
-  it('renders', () => {
-    cy.mount(() => <UseCollapsibleTreeExample/>)
-  })
-})
